requirements.txt
# Project Dependencies

fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
requests
